+++ NEWS.md
@@ -1,6 +1,8 @@
 # parseLatex 0.2.0
 
 * Added `find_sequence()`, `set_contents()`.
+* Add `DEFINITION` tag to allow `\begin` etc. to be temporarily
+ignored.
 
 # parseLatex 0.1.0